Access Control

+++ routes/articles.js
@@ -3,9 +3,10 @@ const router = express.Router();
 
 // Bring in models
 let Articles = require('../models/articles');
+let User = require('../models/user');
 
 // Add route
-router.get('/add', (req, res) => {
+router.get('/add', ensureAuthenticate, (req, res) => {
     res.render('add_article', {
         title: 'Add Articles'
     });
@@ -14,7 +15,7 @@ router.get('/add', (req, res) => {
 // Add submit POST route
 router.post('/add', (req, res) => {
     req.checkBody('title', 'Title is required').notEmpty();
-    req.checkBody('author', 'Author is required').notEmpty();
+    //req.checkBody('author', 'Author is required').notEmpty();
     req.checkBody('body', 'Body is required').notEmpty();
 
     // Get Errors
@@ -28,7 +29,7 @@ router.post('/add', (req, res) => {
     } else {
         let article = new Articles();
         article.title = req.body.title;
-        article.author = req.body.author;
+        article.author = req.user._id;
         article.body = req.body.body;
 
         article.save((err) => {
@@ -44,8 +45,12 @@ router.post('/add', (req, res) => {
 });
 
 // Load Edit Form
-router.get('/edit/:id', (req, res) => {
+router.get('/edit/:id', ensureAuthenticate, (req, res) => {
     Articles.findById(req.params.id, (err, article) => {
+        if(article.author != req.user._id) {
+            req.flash('danger', 'Not Authorized');
+            res.redirect('/');
+        }
         if(err) {
             console.log(err);
         } else {
@@ -79,27 +84,51 @@ router.post('/edit/:id', (req, res) => {
 
 // Delete submit POST route
 router.delete('/:id', (req,res) => {
+
+    if(!req.user._id) {
+        res.status(500).send();
+    }
+
     let query = {_id: req.params.id};
 
-    Articles.deleteOne(query, (err) => {
-        if(err) {
-            console.log(err);
+    Articles.findById(req.params.id, (err, article) => {
+        if(article.author != req.user._id) {
+            res.status(500).send();
+        } else {
+            Articles.deleteOne(query, (err) => {
+                if(err) {
+                    console.log(err);
+                }
+            res.send('Success');
+            });
         }
-    res.send('Success');
     });
 });
 
 // Get Single article
 router.get('/:id', (req, res) => {
     Articles.findById(req.params.id, (err, article) => {
-        if(err) {
-            console.log(err);
-        } else {
-            res.render('article', {
-                article: article
-            });
-        }
+        User.findById(article.author, (err, user) => {
+            if(err) {
+                console.log(err);
+            } else {
+                res.render('article', {
+                    article : article,
+                    author : user.name
+                });
+            }    
+        });
     });
 });
+
+// Access Control
+function ensureAuthenticate(req, res, next) {
+    if(req.isAuthenticated()){
+        return next();
+    } else {
+        req.flash('danger', 'Please login');
+        res.redirect('/users/login');
+    }
+}
 
 module.exports = router;
